requirements.txt

fastapi
uvicorn[standard]

pydantic-settings
python-dotenv


chromadb
aiohttp

loguru
duckduckgo-search
googlesearch-python